Pull out PostgresEventWatcher into its own module and parameterize it more
