+++ python_modules/libraries/dagster-postgres/dagster_postgres/event_log/__init__.py
@@ -1,1 +1,2 @@
 from .event_log import PostgresEventLogStorage
+from .event_watcher import PostgresEventWatcher
